Modified CandidateOneWayConverter, and fixed converters configuration.
